# docs/swagger.yaml
basePath: /gateway/v1
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        description: Email and Phone only one can be set
        example: user1@example.com
        type: string
      name:
        example: user1
        maxLength: 32
        minLength: 2
        type: string
      password:
        example: "123456"
        maxLength: 20
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
    required:
    - name
    - password
    type: object
...
  dto.UpdateUserRequest:
    properties:
      avatar:
        example: https://www.example.com/avatar.png
        type: string
      email:
        example: user1@example.com
        type: string
      name:
        example: user1
        maxLength: 32
        minLength: 2
        type: string
      password:
        example: "123456"
        maxLength: 20
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
    type: object
  dto.User:
    properties:
      avatar:
        example: https://www.example.com/avatar.png
        type: string
      connectUrl:
        example: ws://10.0.0.1:8080/ws
        type: string
      email:
        example: abc@example.com
        type: string
      loginStatus:
        example: 0
        type: integer
      name:
        example: user1
        type: string
      phone:
        example: "13800138000"
        type: string
      uid:
        example: av8FMdRdcb
        type: string
    type: object
  dto.UserLoginRequest:
    properties:
      email:
        description: Email and Phone only one can be set
        example: user1@example.com
        type: string
      loginType:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      password:
        example: "123456"
        maxLength: 20
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
    required:
    - loginType
    - password
    type: object
  response.Meta:
    properties:
      extra:
        additionalProperties:
          type: string
        type: object
      page:
        example: 1
        minimum: 1
        type: integer
      pageSize:
        example: 10
        minimum: 1
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      reason:
        type: string
    type: object
info:
  contact:
    email: yusankurban@gmail.com
    name: Yusank
    url: https://yusank.space
  description: GoIM.Gateway 服务器接口文档
  license:
    name: MIT
    url: https://github.com/go-goim/core/blob/main/LICENSE
  termsOfService: http://go-goim.github.io/
  title: GoIM.Gateway Swagger
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.User'
              type: object
      summary: 登录
      tags:
      - 用户
  /user/query:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 查询用户信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: email
        in: query
        name: email
        type: string
      - description: phone
        in: query
        name: Phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询用户信息
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 注册
      tags:
      - 用户
  /user/update:
    post:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新用户信息
      tags:
      - 用户
swagger: "2.0"